Use shorter datatype for IR time

+++ ir.h
@@ -7,7 +7,7 @@
 typedef uint32_t ir_payload_t;
 
 // IR time datatype
-typedef int16_t ir_time_t;
+typedef uint8_t ir_time_t;
 
 // IR buffer size datatype
 typedef uint8_t ir_size_t;
